Prepared methods for the exam
Made a method that runs once for each input file.
Removed trimming of arrays, as they are not
large enough for it to matter.

Arrays for input data are now initialized for each time
initialize() is called.

Made processing of input data lines, check value of counter for
shortcut.

Uncommented methods for sorting input arrays,
and made them use the set() method, which saves a lot of time...
(from 800 microseconds to 30 microseconds for input file 5)

Getters now return List instead of ArrayList.

Improved the output text formatting.