Can now find start index of a vehicle; and find all empty vehicles
Added method that finds the start index of a vehicle,
given the last index or an index within the vehicle.

Added a method that finds all the vehicles that are not assigned
any calls, and return their start indices in a list.

Also added parameter checks for startIndex() and stopIndex()